+++ src/main/kotlin/blackjack/MainApp.kt
@@ -1,11 +1,19 @@
 package blackjack
 
-import view.InputView
-import view.ResultView
+import blackjack.domain.Game
+import blackjack.view.InputView
+import blackjack.view.OutputView
 
 fun main() {
     val playerNames = InputView.readSetup()
     val game = Game(playerNames)
     val players = game.allPlayers()
-    ResultView.showResultOfSetUp(players)
+    OutputView.showResultOfSetUp(players)
+    while (!game.isOver()) {
+        val currentPlayer = game.currentlyPlayer()
+        val reply = InputView.replyDraw(currentPlayer)
+        val playerAfterDrawing = game.chanceDraw(reply)
+        OutputView.showStateOfCards(playerAfterDrawing!!, reply)
+    }
+    OutputView.showResult(game.result())
 }
