+++ src/main/kotlin/blackjack/view/InputView.kt
@@ -1,6 +1,6 @@
-package view
+package blackjack.view
 
-import blackjack.Player
+import blackjack.domain.Player
 
 const val ENTER_PLAYERS_NAMES = "게임에 참여할 사람의 이름을 입력하세요.(쉼표 기준으로 분리)"
 const val CHOOSE_DRAW_OR_NOT = "는 한 장의 카드를 더 받겠습니까? (예는 y, 아니오는 n)"
@@ -20,9 +20,8 @@ object InputView {
         return readText(ENTER_PLAYERS_NAMES)
     }
 
-    fun chooseDraw(player: Player): String {
-        var text = readText("\n" + player + CHOOSE_DRAW_OR_NOT)
-        val availableReplies = listOf("y", "n")
+    fun replyDraw(player: Player): String {
+        var text = readText("\n$player"+CHOOSE_DRAW_OR_NOT)
         return text
     }
 }
