+++ src/test/kotlin/blackjack/DenominationTest.kt
@@ -1,5 +1,6 @@
 package blackjack
 
+import blackjack.domain.Denomination
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
